+++ matomo.php
@@ -4,10 +4,10 @@
  * Description: Privacy friendly, GDPR compliant and self-hosted. Matomo is the #1 Google Analytics alternative that gives you control of your data. Free and secure.
  * Author: Matomo
  * Author URI: https://matomo.org
- * Version: 5.1.5
+ * Version: 5.1.6
  * Domain Path: /languages
  * WC requires at least: 2.4.0
- * WC tested up to: 9.3.3
+ * WC tested up to: 9.4.1
  *
  * Matomo - free/libre analytics platform
  *