+++ tests/phpunit/wpmatomo/wpstatistics/test-import.php
@@ -39,7 +39,7 @@ class ImportTest extends MatomoAnalytics_TestCase {
 		if ( file_exists( $file ) ) {
 			require_once $file;
 
-			$wp_statistics = $GLOBALS['WP_Statistics'];
+			$wp_statistics = \WP_Statistics();
 			if ( method_exists( $wp_statistics, 'plugin_setup' ) ) {
 				$wp_statistics->plugin_setup();
 			} else {
@@ -56,6 +56,8 @@ class ImportTest extends MatomoAnalytics_TestCase {
 			global $wp_filesystem;
 
 			require_once $this->plugin_file();
+
+			set_current_screen( 'edit-post' ); // so is_admin() will return true
 
 			$this->data = new Data();
 
@@ -76,11 +78,17 @@ class ImportTest extends MatomoAnalytics_TestCase {
 				$this->download_geoip();
 				$this->manually_load_plugin();
 			}
+
+			// update the wp-statistics database
+			\WP_STATISTICS\Install::create_table( is_multisite() );
+			\WP_STATISTICS\Install::create_options();
+
 			// run the import
 			$importer = new Importer( new \Psr\Log\NullLogger() );
 			$site     = new Site();
 			$id_site  = $site->get_current_matomo_site_id();
 			// do not run the archiving for performances issues and because we test only daily reports
+			$importer->set_should_rethrow( true );
 			$importer->import( $id_site, false );
 		}
 	}
@@ -104,8 +112,12 @@ class ImportTest extends MatomoAnalytics_TestCase {
 		}
 
 		$wpstats_database_path = ABSPATH . '/wp-content/uploads/wp-statistics/GeoLite2-City.mmdb';
-		mkdir( dirname( $wpstats_database_path ), 0777, true );
-		symlink( $expected_path, $wpstats_database_path );
+		if ( ! is_dir( dirname( $wpstats_database_path ) ) ) {
+			mkdir( dirname( $wpstats_database_path ), 0777, true );
+		}
+		if ( ! is_file( $wpstats_database_path ) ) {
+			symlink( $expected_path, $wpstats_database_path );
+		}
 	}
 
 	public function test_countries_found() {
@@ -139,7 +151,7 @@ class ImportTest extends MatomoAnalytics_TestCase {
 
 		$report = $this->fetch_report( 'UserCountry', 'getCity' );
 		// 500 due to the limit in the datatable
-		$this->assertEquals( $report['reportData']->getRowsCount(), 500 );
+		$this->assertEquals( 500, $report['reportData']->getRowsCount() );
 	}
 
 	public function test_browsers_found() {
@@ -150,7 +162,7 @@ class ImportTest extends MatomoAnalytics_TestCase {
 		}
 
 		$report = $this->fetch_report( 'DevicesDetection', 'getBrowsers' );
-		$this->assertEquals( $report['reportData']->getRowsCount(), 15 );
+		$this->assertGreaterThanOrEqual( 15, $report['reportData']->getRowsCount() );
 	}
 
 	public function test_os_found() {
@@ -161,7 +173,7 @@ class ImportTest extends MatomoAnalytics_TestCase {
 		}
 
 		$report = $this->fetch_report( 'DevicesDetection', 'getOsVersions' );
-		$this->assertEquals( $report['reportData']->getRowsCount(), 10 );
+		$this->assertEquals( 10, $report['reportData']->getRowsCount() );
 	}
 
 	public function test_referrers_found() {
@@ -172,7 +184,7 @@ class ImportTest extends MatomoAnalytics_TestCase {
 		}
 
 		$report = $this->fetch_report( 'Referrers', 'getWebsites' );
-		$this->assertEquals( $report['reportData']->getRowsCount(), 49 );
+		$this->assertEquals( 49, $report['reportData']->getRowsCount() );
 	}
 
 	public function test_search_engines_found() {
@@ -183,18 +195,7 @@ class ImportTest extends MatomoAnalytics_TestCase {
 		}
 
 		$report = $this->fetch_report( 'Referrers', 'getSearchEngines' );
-		$this->assertEquals( $report['reportData']->getRowsCount(), 6 );
-	}
-
-	public function test_keywords_found() {
-		if ( ! $this->can_be_tested() ) {
-			$this->markTestSkipped( 'CI or plugin unavailable' );
-
-			return;
-		}
-
-		$report = $this->fetch_report( 'Referrers', 'getKeywords' );
-		$this->assertEquals( $report['reportData']->getRowsCount(), 2 );
+		$this->assertEquals( 6, $report['reportData']->getRowsCount() );
 	}
 
 	public function test_visitors_found() {
@@ -205,7 +206,10 @@ class ImportTest extends MatomoAnalytics_TestCase {
 		}
 
 		$report = $this->fetch_report( 'VisitsSummary', 'get' );
-		$this->assertEquals( $report['reportData']->getFirstRow()->getColumn( 'nb_visits' ), 1298 );
+		$row    = $report['reportData']->getFirstRow();
+
+		$this->assertInstanceOf( \Piwik\DataTable\Row::class, $row );
+		$this->assertEquals( 1298, $row->getColumn( 'nb_visits' ) );
 	}
 
 	public function test_pages_found() {